+++ tests/testthat/test-help.R
@@ -4,17 +4,17 @@ test_that("shim_help behaves the same as utils::help for non-devtools-loaded pac
   # stats wasn't loaded with devtools. There are many combinations of calling
   # with quotes and without; make sure they're the same both ways. Need to index
   # in using [1] to drop attributes for which there are unimportant differences.
-  expect_identical(shim_help(lm)[1],            utils::help(lm)[1])
-  expect_identical(shim_help(lm, stats)[1],     utils::help(lm, stats)[1])
-  expect_identical(shim_help(lm, 'stats')[1],   utils::help(lm, 'stats')[1])
-  expect_identical(shim_help('lm')[1],          utils::help('lm')[1])
-  expect_identical(shim_help('lm', stats)[1],   utils::help('lm', stats)[1])
+  expect_identical(shim_help(lm)[1], utils::help(lm)[1])
+  expect_identical(shim_help(lm, stats)[1], utils::help(lm, stats)[1])
+  expect_identical(shim_help(lm, 'stats')[1], utils::help(lm, 'stats')[1])
+  expect_identical(shim_help('lm')[1], utils::help('lm')[1])
+  expect_identical(shim_help('lm', stats)[1], utils::help('lm', stats)[1])
   expect_identical(shim_help('lm', 'stats')[1], utils::help('lm', 'stats')[1])
-  expect_identical(shim_help(, "stats")[1],     utils::help(, "stats")[1])
+  expect_identical(shim_help(, "stats")[1], utils::help(, "stats")[1])
 
   # Works for :: and ::: as well (#72)
-  expect_identical(shim_help("::")[1],          utils::help("::")[1])
-  expect_identical(shim_help(":::")[1],         utils::help(":::")[1])
+  expect_identical(shim_help("::")[1], utils::help("::")[1])
+  expect_identical(shim_help(":::")[1], utils::help(":::")[1])
 })
 
 test_that("shim_help behaves the same as utils::help for nonexistent objects", {
@@ -91,8 +91,14 @@ test_that("shim_help and shim_questions works if topic moves", {
   expect_equal(base_rd_path(shim_help("foofoo")), "foofoo.Rd")
   expect_equal(base_rd_path(shim_question("foofoo")), "foofoo.Rd")
 
-  fs::file_move(fs::path(path_man, "foofoo.Rd"), fs::path(path_man, "barbar.Rd"))
-  defer(fs::file_move(fs::path(path_man, "barbar.Rd"), fs::path(path_man, "foofoo.Rd")))
+  fs::file_move(
+    fs::path(path_man, "foofoo.Rd"),
+    fs::path(path_man, "barbar.Rd")
+  )
+  defer(fs::file_move(
+    fs::path(path_man, "barbar.Rd"),
+    fs::path(path_man, "foofoo.Rd")
+  ))
 
   expect_equal(base_rd_path(shim_help("foofoo")), "barbar.Rd")
   expect_equal(base_rd_path(shim_question("foofoo")), "barbar.Rd")
@@ -133,7 +139,13 @@ test_that("unknown macros don't trigger warnings (#119)", {
 
 test_that("complex expressions are checked", {
   expect_snapshot({
-    (expect_error(shim_help({ foo; bar }), "must be a name"))
+    (expect_error(
+      shim_help({
+        foo
+        bar
+      }),
+      "must be a name"
+    ))
   })
 })