+++ tests/testthat/test-load-hooks.R
@@ -13,30 +13,30 @@ test_that("hooks called in correct order", {
     h$events <- character()
   }
 
-
   setHook(packageEvent("testHooks", "attach"), record_use("user_attach"))
   setHook(packageEvent("testHooks", "detach"), record_use("user_detach"))
-  setHook(packageEvent("testHooks", "onLoad"),   record_use("user_load"))
+  setHook(packageEvent("testHooks", "onLoad"), record_use("user_load"))
   setHook(packageEvent("testHooks", "onUnload"), record_use("user_unload"))
 
   reset_events()
   load_all("testHooks")
-  expect_equal(globalenv()$hooks$events,
+  expect_equal(
+    globalenv()$hooks$events,
     c("pkg_load", "user_load", "pkg_attach", "user_attach")
   )
 
   reset_events()
   unload("testHooks")
-  expect_equal(globalenv()$hooks$events,
+  expect_equal(
+    globalenv()$hooks$events,
     c("user_detach", "pkg_detach", "user_unload", "pkg_unload")
   )
 
   rm(list = "hooks", envir = globalenv())
   setHook(packageEvent("testHooks", "attach"), NULL, "replace")
   setHook(packageEvent("testHooks", "detach"), NULL, "replace")
-  setHook(packageEvent("testHooks", "onLoad"),   NULL, "replace")
+  setHook(packageEvent("testHooks", "onLoad"), NULL, "replace")
   setHook(packageEvent("testHooks", "onUnload"), NULL, "replace")
-
 })
 
 test_that("onLoad and onAttach", {