+++ tests/testthat/test-s4-export.R
@@ -3,7 +3,7 @@ local_load_all_quiet()
 test_that("importing an S4 exported by another pkg with export_all = FALSE", {
   load_all("testS4export", export_all = FALSE)
 
-  # this used to crash with error: 
+  # this used to crash with error:
   # class "class_to_export" is not exported by 'namespace:testS4export'
   load_all("testS4import", export_all = FALSE)
   expect_true(isClassDef(getClass("derived")))