+++ tests/testthat/testActiveBindings/R/bindings.r
@@ -1,4 +1,3 @@
-
 makeActiveBinding("foo", function() rlang::abort("foo"), environment())
 
 #' @export